# GH-36176: [C++] Fix regression for single-key Table sorting