@pitrou 2023-06-20
This fixes the regression here.